scheduled_jobs/legacy/raw/requirements.txt
# Function dependencies, for example:
# package>=version
requests
mysql-connector-python
snowflake-connector-python